Refactor code to use a custom hook.
