+++ assets/src/edit-story/app/story/actions/useSavePost.js
@@ -1,0 +1,34 @@
+/**
+ * WordPress dependencies
+ */
+import { useCallback } from '@wordpress/element';
+/**
+ * Internal dependencies
+ */
+import { useAPI } from '../../api';
+
+function useSavePost( {
+	isSaving,
+	storyId,
+	title,
+	postStatus,
+	pages,
+	setPostStatus,
+	setIsSaving,
+} ) {
+	const status = ( postStatus !== 'publish' ) ? 'publish' : postStatus;
+	const { actions: { saveStoryById } } = useAPI();
+	const savePost = useCallback( () => {
+		if ( ! isSaving ) {
+			setIsSaving( true );
+			saveStoryById( storyId, title, status, pages ).then( () => {
+				setIsSaving( false );
+				setPostStatus( status );
+			} );
+		}
+	}, [ isSaving, setIsSaving, storyId, title, pages, setPostStatus ] );
+
+	return savePost;
+}
+
+export default useSavePost;
